vyos.ifconfig: T5103: always stop the DHCP client process bevore changing VRF
Always stop the DHCP client process to clean up routes within the VRF where the
process was originally started. There is no need to add a condition to only
call the method if "address dhcp" was defined, as this is handled inside
set_dhcp(v6) by only stopping if the daemon is running.

DHCP client process restart will be handled later on once the interface is
moved to the new VRF.