T5400: add local build of libvyosconfig to Makefile
libvyosconfig is both a build and a run dependency of vyos-1x.
Satisfying the build dependency within the Docker image requires
coordination of updates to vyos-build/libvyosconfig/vyos-1x on any
changes to the library; simplify this process by moving the build to a
step of the vyos-1x Makefile.

(cherry picked from commit 731195c7adb21208767199f5487dfdf29ce09380)