pki: T7122: when ACME listen-address is used - check if port is available
When instructing certbot to listen on a given address, check if the address is
free to use. Also take this into account when spawning certbot behind HAProxy.
If the address is not (yet) bound - the request must be done in standalone mode
and not via the reverse-proxy.

(cherry picked from commit f40cf6064a02fbb6baae924e94b9183d6bd87474)