T7510: add commit warnings about invalid use of OSPF area-types
To keep existing CLI behavior use a Warning() to prompt the user for an invalid
configuration. It is not possible to have more the one area-type defined per
area logically - the CLI does support it. In addition the backbone area cannot
be of type STUB or NSSA.

CLI configuration should be cleaned up using a migrator in the future.

(cherry picked from commit 323ef231bd06650de1405daf6d04b59c1b7b4285)